* base/drop_privileges.c (drop_privileges): Drop supplementary groups 	too.
	* base/drop_privileges.h: Add define accordingly.

git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/openvas-libraries@21903 423fd1db-d629-0410-8442-d21db03e70f4

+++ base/drop_privileges.c
@@ -38,7 +38,8 @@
 
 #include <pwd.h>
 #include <unistd.h>
-
+#include <sys/types.h>
+#include <grp.h>
 
 /**
  * @brief Sets an error and return \param errorcode.
@@ -89,14 +90,21 @@ drop_privileges (gchar * username, GError ** error)
     {
       if ((user_pw = getpwnam (username)))
         {
+          if (initgroups (username, user_pw->pw_gid) != 0)
+            return drop_privileges_error
+                    (error,
+                     OPENVAS_DROP_PRIVILEGES_FAIL_SUPPLEMENTARY,
+                     "Failed to drop supplementary groups privileges!\n");
           if (setgid (user_pw->pw_gid) != 0)
-            return drop_privileges_error (error,
-                                          OPENVAS_DROP_PRIVILEGES_FAIL_DROP_GID,
-                                          "Failed to drop group privileges!\n");
+            return drop_privileges_error
+                    (error,
+                     OPENVAS_DROP_PRIVILEGES_FAIL_DROP_GID,
+                     "Failed to drop group privileges!\n");
           if (setuid (user_pw->pw_uid) != 0)
-            return drop_privileges_error (error,
-                                          OPENVAS_DROP_PRIVILEGES_FAIL_DROP_UID,
-                                          "Failed to drop user privileges!\n");
+            return drop_privileges_error
+                    (error,
+                     OPENVAS_DROP_PRIVILEGES_FAIL_DROP_UID,
+                     "Failed to drop user privileges!\n");
         }
       else
         {
